fix: update href, Heroicons libs, fix import wrong, change filename
