+++ src/components/index.ts
@@ -4,7 +4,9 @@ import ButtonOutline from './Button/ButtonOutline';
 import ButtonSolid from './Button/ButtonSolid';
 import Input from './Input/Input';
 import InputRadio from './Input/InputRadio';
+import DropListSV from '@/components/DropList_SV/DropListSV';
 
 export { classNames };
 export { ButtonBase, ButtonSolid, ButtonOutline };
 export { Input, InputRadio };
+export { DropListSV };
